requirements.txt
Flask==3.0.3
Pillow==10.4.0
pyzbar==0.1.9